Move validation logic into its own service
This allows reusability in something like a Drush command and
dramatically cleans up the logic of the controller.

While doing this we also add a simple passthrough to validate the schema
which is also useful to show users.
