+++ modules/custom/social_graphql/modules/social_graphql_devel/social_graphql_devel.links.task.yml
@@ -1,4 +1,4 @@
-social_graphql_devel.validate_resolvers:
-  route_name: social_graphql_devel.validate_resolvers
+social_graphql_devel.validate:
+  route_name: social_graphql_devel.validate
   base_route: entity.graphql_server.edit_form
-  title: Validate Resolvers
+  title: Validate
