+++ modules/custom/social_graphql/modules/social_graphql_devel/social_graphql_devel.routing.yml
@@ -1,8 +1,8 @@
-social_graphql_devel.validate_resolvers:
-  path: '/admin/config/graphql/servers/manage/{graphql_server}/validate_resolvers'
+social_graphql_devel.validate:
+  path: '/admin/config/graphql/servers/manage/{graphql_server}/validate'
   defaults:
-    _controller: '\Drupal\social_graphql_devel\Controller\ValidateResolversController::validateResolvers'
-    _title: 'Validate Resolvers'
+    _controller: '\Drupal\social_graphql_devel\Controller\ValidationController::report'
+    _title: 'Validate GraphQL Server'
   requirements:
     _permission: 'administer graphql configuration'
   options:
